refactor(schemas): move legacy systemMigrations/contentSchemaEvents to _legacy/ (drop in Plan B)
Per-dialect Drizzle table definitions for systemMigrations and
contentSchemaEvents now live in schemas/_legacy/{postgres,mysql,sqlite}.ts
with a dialect-namespaced barrel (`pg`, `my`, `sl` re-exports). Each file
carries an @deprecated JSDoc block flagging these tables for removal in
Plan B's first task.

database/schema/<dialect>.ts re-exports from the new location so existing
imports keep compiling until Task 16 sweeps imports to @nextly/schemas
and Plan B drops the tables entirely.

Note: contentSchemaEvents is superseded by nextly_schema_events;
systemMigrations was the legacy ledger replaced by the new
domains/schema/pipeline. The entire _legacy/ directory is deleted by Plan
B's first task.
